Assignment7/faiss_index/requirements.txt
faiss-cpu>=1.7.4
flask>=2.0.0
flask-cors>=3.0.10
numpy>=1.23.0
beautifulsoup4>=4.11.0
python-dotenv>=1.0.0
requests>=2.28.0